Switch interface to comet driven updates.
The backend checks are not fed to the interface using the comet protocol.
This allows backend checks to take more time to complete without hindering
the responsiveness of the interface.

No progress bar is currently implemented, but backend checks can specify
an estimated runtime so that they will be factored into a progress bar
correctly once it is implemented. The default runtime for a check when
none is specified is 1.

The interface now provides a button to rerun the scan again.