Corrected kernel version detection
Case 110893: Utilize the Cpanel::Kernel module to determine which kernel
should be booted into on the next restart.  If we are not currently
running that kernel version, throw a warning.

This change assumes that any newly installed kernels will automatically
update the grub configuration file and that only the latest desired kernel
version will be selected for boot.  If administrators install kernel
versions out of order, this could potentially report incorrectly.

This was done to avoid issues when multiple types of kernels are
installed, e.g. CloudLinux and CentOS.  Switching to CentOS might bring in
a lower number kernel, but we want to make sure we boot into a CentOS
kernel -- the type we just installed -- not the CloudLinux one.

This change relies on Cpanel::Kernel, which has some limitations.
Firstly, this will not work when systems utilize a "saved" default boot in
their GRUB configuration, which can be used to setup fallback or once-only
boots.  Secondly, it's limited to GRUB Legacy.