Fix detection of CloudLinux kernels.
Case CPANEL-3865: The CloudLinux kernel's lve module changed its name to
kmodlve.  Look for either one of these module names.

Note that querying /etc/redhat-release is not enough to determine
whether the kernel is a CloudLinux kernel, since it's possible to run
CloudLinux on a CentOS kernel, which can happen (usually by accident) in
some configurations.