Avoid logging error messages in MySQL test.
Case CPANEL-5699: An error message would appear in the cPanel error log
when running the MySQL test because the command to check for anonymous
users failed.  Since that failure is desirable behavior (as it means the
system is secure), add an option to silence that output.  With a
corresponding change in 11.58, this allows us to not log errors in this
particular case.